implement Optim as a weak dependency
