+++ Project.toml
@@ -13,6 +13,12 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 
+[weakdeps]
+Optim = "429524aa-4258-5aef-a3af-852621145aeb"
+
+[extensions]
+DistributionFitsOptimExt = "Optim"
+
 [compat]
 Distributions = "0.25"
 FillArrays = "0.12, 0.13"
@@ -22,3 +28,5 @@ StaticArrays = "1.2"
 StatsBase = "0.33"
 StatsFuns = "0.9, 1"
 julia = "1.6"
+Optim = "1.7"
+
