+++ test/runtests.jl
@@ -3,16 +3,17 @@ using Test
 using Random: Random
 
 @testset "optimize error" begin
-    @test_throws Exception DistributionFits.optimize(x -> x*x, DistributionFits.optimizer, -1, 1)
+    @test_throws Exception DistributionFits.optimize(x -> x*x, -1, 1)
 end
 # Optim package for interactive testing
 i_loadlibs = () -> begin
     push!(LOAD_PATH, expanduser("~/julia/scimltools/")) # access local package repo
+    push!(LOAD_PATH, expanduser("~/julia/18_tools/scimltools/")) # access local package repo
 end
 using Optim: Optim, optimize
 @testset "optimize set in __init__ after using Optim" begin
     # set in __init__
-    @test DistributionFits.optimizer isa DistributionFits.OptimOptimizer
+    @test DistributionFits.df_optimizer isa DistributionFits.DistributionFitsOptimExt.OptimOptimizer
 end
 
 const tests = [
